Clean up includes & u8 -> u16 realmID

+++ src/ECS/Components/Network/LoadBalanceSingleton.h
@@ -3,13 +3,14 @@
 #include <Networking/NetStructures.h>
 #include <entity/fwd.hpp>
 #include <vector>
+#include <robin_hood.h>
 
 #pragma pack(push, 1)
 struct ServerInformation
 {
     entt::entity entity = entt::null;
     AddressType type = AddressType::INVALID;
-    u8 realmId = 0;
+    u16 realmId = 0;
     u32 address = 0;
     u16 port = 0;
 };
@@ -198,7 +199,7 @@ struct LoadBalanceSingleton
             if (numOf == 0)
                 return false;
 
-            u8& index = realmServersIndex[realmId];
+            u16& index = realmServersIndex[realmId];
             serverInformation = realmServersMap[realmId][index++];
 
             // Wrap index around if needed
@@ -211,7 +212,7 @@ struct LoadBalanceSingleton
             if (numOf == 0)
                 return false;
 
-            u8& index = worldServersIndex[realmId];
+            u16& index = worldServersIndex[realmId];
             serverInformation = worldServersMap[realmId][index++];
 
             // Wrap index around if needed
@@ -224,7 +225,7 @@ struct LoadBalanceSingleton
             if (numOf == 0)
                 return false;
 
-            u8& index = instanceServersIndex[realmId];
+            u16& index = instanceServersIndex[realmId];
             serverInformation = instanceServersMap[realmId][index++];
 
             // Wrap index around if needed
@@ -236,21 +237,21 @@ struct LoadBalanceSingleton
     }
 
 private:
-    u8 authIndex = 0;
-    u8 loadBalanceIndex = 0;
-    u8 regionIndex = 0;
-    u8 chatIndex = 0;
+    u16 authIndex = 0;
+    u16 loadBalanceIndex = 0;
+    u16 regionIndex = 0;
+    u16 chatIndex = 0;
 
     std::vector<ServerInformation> authServers;
     std::vector<ServerInformation> loadBalancers;
     std::vector<ServerInformation> regionServers;
     std::vector<ServerInformation> chatServers;
 
-    std::vector<u8> realmServersIndex;
-    std::vector<u8> worldServersIndex;
-    std::vector<u8> instanceServersIndex;
+    std::vector<u16> realmServersIndex;
+    std::vector<u16> worldServersIndex;
+    std::vector<u16> instanceServersIndex;
 
-    robin_hood::unordered_map<u8, std::vector<ServerInformation>> realmServersMap;
-    robin_hood::unordered_map<u8, std::vector<ServerInformation>> worldServersMap;
-    robin_hood::unordered_map<u8, std::vector<ServerInformation>> instanceServersMap;
+    robin_hood::unordered_map<u16, std::vector<ServerInformation>> realmServersMap;
+    robin_hood::unordered_map<u16, std::vector<ServerInformation>> worldServersMap;
+    robin_hood::unordered_map<u16, std::vector<ServerInformation>> instanceServersMap;
 };
